requirements.txt
python-3.7.0
redis-3.0.1


